[DBScripts] Implement SCRIPT_COMMAND_SEND_MAIL (c2639)
Attention DB-Devs:

The fields quest_template.RewMailTemplateId and .RewMailDelaySecs are scheduled to be removed

+++ src/game/WorldHandlers/ScriptMgr.cpp
@@ -38,6 +38,7 @@
 #include "OutdoorPvP/OutdoorPvP.h"
 #include "WaypointMovementGenerator.h"
 #include "Mail.h"
+//#include "LFGMgr.h"
 
 #ifdef ENABLE_ELUNA
 #include "LuaEngine.h"
@@ -2025,7 +2026,7 @@ void ScriptMgr::LoadScriptBinding()
     {
         if (m_scriptBind[i].size()) //ignore missing script types to shorten the log
         {
-            sLog.outString(".. type %u: %u binds", i, m_scriptBind[i].size());
+            sLog.outString(".. type %u: %u binds", i, uint32(m_scriptBind[i].size()));
             count -= m_scriptBind[i].size();
         }
     }
